refactoring messages and users code
