+++ gymbuddies/Controllers/MessagesController.swift
@@ -1,24 +1,12 @@
 //
-//  RecentMessagesView.swift
+//  MessagesController.swift
 //  gymbuddies
 //
-//  Created by Kareha on 2/3/21.
+//  Created by Kareha on 2/11/21.
 //
 
 import SwiftUI
 import Firebase
-
-struct RecentMessagesView: View {
-    var body: some View {
-        MessagesView().environmentObject(MainObservable())
-    }
-}
-
-struct RecentMessagesView_Previews: PreviewProvider {
-    static var previews: some View {
-        RecentMessagesView()
-    }
-}
 
 class MainObservable : ObservableObject{
     
@@ -90,8 +78,4 @@ class MainObservable : ObservableObject{
             }
         }
     }
-    
 }
-
-
-
