+++ gymbuddies/Views/SignInView.swift
@@ -6,7 +6,7 @@
 //
 
 import SwiftUI
-import Firebaseg
+import Firebase
 
 struct SignInView: View {
